# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for Taskify application.",
        "title": "Taskify API",
        "contact": {},
        "version": "1.0"
    },
    "host": "185.68.22.208:3000",
    "basePath": "/api",
    "paths": {
        "/users": {
            "post": {
                "description": "Создает пользователя, используя поля: email, username, password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Создание пользователя",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "RequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.CreateUserDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/users.User"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Получает пользователя по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Получение пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.User"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновления пользовательских полей: email, username, password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Обновление пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Поля для редактирования",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.CreateUserDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.User"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет пользователя по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Удаление пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.User"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/boards": {
            "post": {
                "description": "Создает пользователя, используя id пользователя и поля: name, description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Создание таблицы",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для создания",
                        "name": "RequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/boards.CreateBoardDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/boards.Board"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/boards/{id}": {
            "get": {
                "description": "Получает таблицы пользователя по ее порядковому номеру",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Получение таблицы",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id таблицы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/boards.Board"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновления полей таблицы: name, description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Обновление таблицы",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id таблицы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Поля для редактирования",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/boards.CreateBoardDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/boards.Board"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет таблицу пользователя по ее порядковому номеру",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Удаление таблицы",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id таблицы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/boards.Board"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "boards.Board": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Deposit feature development"
                },
                "id": {
                    "type": "integer",
                    "example": 101
                },
                "name": {
                    "type": "string",
                    "example": "Deposit"
                }
            }
        },
        "boards.CreateBoardDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Deposit feature development"
                },
                "name": {
                    "type": "string",
                    "example": "Deposit"
                }
            }
        },
        "users.CreateUserDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "taskify@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                },
                "username": {
                    "type": "string",
                    "example": "johnlennon"
                }
            }
        },
        "users.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "taskify@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 101
                },
                "username": {
                    "type": "string",
                    "example": "johnlennon"
                }
            }
        }
    }
}